2016-04-25 Roman - DAQ: Initialization of redundancy of TController object some changed to enabled by default. That will next checked into a complex environment. - DAQ.BCM2835: Expanded for access to GPIO by static functions which significant faster ones. - FIX:DAQ.JavaLikeCalc: setNoTransl() at the attributes value (DB field "DEF") loading disabled by some problematic and no a sence. - Special.FLibSYS: The time functions are appended by tmSleep() for faster access to the short sleeping provides. - Transport.{SSL,Sockets}: Buffer's size of getservbyname() some decreased. - FIX:UI.QTCfg: Read only of selected items are fixed in alignment with other items in single row.     Changed user apply is fixed.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2361 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tsys.cpp
@@ -249,7 +249,7 @@ string TSYS::uint2str( unsigned val, IntView view )
     return buf;
 }
 
-string TSYS::ll2str( int64_t val, IntView view )
+string TSYS::ll2str( long long val, IntView view )
 {
     char buf[NSTR_BUF_LEN];
     switch(view) {
@@ -2176,7 +2176,7 @@ TVariant TSYS::objFuncCall( const string &iid, vector<TVariant> &prms, const str
 	return string("0");
     }
     // int sleep(real tm, int ntm = 0) - call for task sleep to <tm> seconds and <ntm> nanoseconds.
-    //  tm - wait time in seconds, no more to 100 seconds
+    //  tm - wait time in seconds (precised up to nanoseconds), up to STD_INTERF_TM(5 seconds)
     //  ntm - wait time part in nanoseconds
     if(iid == "sleep" && prms.size() >= 1) {
 	struct timespec sp_tm;
