+++ src/tsys.h
@@ -215,7 +215,7 @@ class TSYS : public TCntrNode
 	// Convert value to string
 	static string int2str( int val, IntView view = Dec );
 	static string uint2str( unsigned val, IntView view = Dec );
-	static string ll2str( int64_t val, IntView view = Dec );
+	static string ll2str( long long val, IntView view = Dec );
 	static string real2str( double val, int prec = 15, char tp = 'g' );
 	static double realRound( double val, int dig = 0, bool toint = false ) {
 	    double rez = floor(val*pow(10,dig)+0.5)/pow(10,dig);
